% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary_report.R
\name{create_glm_summary_report}
\alias{create_glm_summary_report}
\title{summary_report}
\usage{
create_glm_summary_report(summary)
}
\arguments{
\item{summary}{summary of a model}
}
\value{
The return value is a list of a ceofficient dataframe, a statistic
indicators dataframe and the formula as a character
}
\description{
A fct function that create a list to make  easier
the exportation of the summary inforamtion
}
