% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dataset_creation.R
\name{dataset_creation}
\alias{dataset_creation}
\title{dataset_creation}
\usage{
dataset_creation(
  start_year = 2000,
  end_year = 2030,
  nb_year = 9,
  start_lat = 44,
  start_long = -2
)
}
\arguments{
\item{start_year}{numeric corresponding to the year of initial statement}

\item{end_year}{numeric corresponding to the end of the values series}

\item{nb_year}{numeric corresponding to total number of year including survey}

\item{start_lat}{numeric : latitude of the corner of the impact area}

\item{start_long}{numeric : longitude of the corner of the impact area}
}
\value{
A list composed of 2 dataframe and 1 polygon.
}
\description{
A fct function
  #creation of a fictional dataset of granulate concession following
there are 3 species found
a list of 3 data frame, 3 shape file (shp, shx, dbf) :
 data_catch give information of the sample
 data_operation give information of the concession
 save_input give information about the ref and impact stations
 shape files give the delimitation of the concession
 don't care between continuous or discrete distribution because on boat
 number is an estimation and sometimes there is relative number


 The number of species per (station,season,year) is determined randomly
 by probability distribution :
species 1 no effect####
rnbinom(720, mu=100, size = 1.1)
####
species 2 processing effect####
 less variance and it's possible to have several 0
 rnbinom(1, mu=20, size = 3)
to model the effect when process we keep only value < 15
rnbinom(1, mu=20, size = 3) < 15
####
species 3  season effect and interaction processing*season####
less variance and it's possible to have several 0
rnbinom(1, mu=100, size = 1)
the gamma have high variance, no 0 and high value
it is easiest to control an increase than with the lnorm
if "summer" = rgamma(1, shape = 2, scale = 300)
if "winter" = rgamma(1, shape = 2, scale = 50)
#same for spring and autumn = rgamma(1, shape = 2, scale = 150)
####
species 4 no effect but big variation pop dynamics
rnbinom(1, mu=100, size = 0.1)

Some parameters needed in the standard format of the app have been fixed
the distance of one station sample is fixed at 1.2e-5
the impact area size is fixed to polygon of 1 latitude per 1 longitude
the date use the format %d%m%Y %H%M and the survey are (01/01, 01/04, 01/07,
 01/10) of each year

There are 20 stations with 10 randomly choose in the impact area and
10 randomly choose outside the impact zone but inside the sample zone
}
