% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_delete_outliers.R
\name{delete_outliers}
\alias{delete_outliers}
\title{delete_outliers}
\usage{
delete_outliers(vector)
}
\arguments{
\item{vector}{a numeric or integer vector}
}
\value{
The intitial vector without the outliers of the series
}
\description{
A fct function that remove the outliers of a vector using
the 1.5xIQR rule
}
