% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_diversity_table.R
\name{diversity_table}
\alias{diversity_table}
\title{diversity_table}
\usage{
diversity_table(data, var_name)
}
\arguments{
\item{data}{dataframe}

\item{var_name}{character of 1 value}
}
\value{
The return value is a dataframe with the value of chosen variable
for interior, exterior of the concession and the globale average
}
\description{
A fct function that create a dataframe compose of 5 indictors
with 3 modalities each : (Abun,Biom,Richness,Shannon,Simpson) ; (int,ext,tot)
}
