% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export_data.R
\name{export_data}
\alias{export_data}
\title{export_data}
\usage{
export_data(path, catch, operation, impact_polygon)
}
\arguments{
\item{path}{string of th folder path}

\item{catch}{dataframe}

\item{operation}{dataframe}

\item{impact_polygon}{sf polygon}
}
\value{
No return value
}
\description{
A fct function to export the rds data of the package in good
format for the app
}
