% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm_maker.R
\name{glm_maker}
\alias{glm_maker}
\title{glm_maker}
\usage{
glm_maker(
  data,
  formule,
  formule_bis,
  language,
  language_bis,
  interaction,
  family
)
}
\value{
The return value is a list of 2 result from the glmm initial or
final model (with or without interaction for example)
}
\description{
Function to run the glm according to the formula created
by the fct ecriture_modele
Interaction is automatically removed if not significant

In the case of glm, you need to modify the formula and the distribution
directly in the glm output object so that it is displayed correctly
in the model's output.
}
