% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glmm_maker.R
\name{glmm_maker}
\alias{glmm_maker}
\title{glmm_maker}
\usage{
glmm_maker(data, formule, formule_bis, interaction, family)
}
\value{
The return value is a list of 2 result from the glmm initial or
final model (with or without interaction for example)
}
\description{
Function to run the glmm according to the formula created
by the fct ecriture_modele
Interaction is automatically removed if not significant

optimisation for better convergence: authorisation of partial derivative
approximations and nAGQ = 0
to be modified as required and as the methodology evolves
}
