% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_function.R
\name{global_function}
\alias{global_function}
\title{global_function}
\usage{
global_function(
  tutti_catch,
  tutti_operation,
  liste_station,
  liste_dates,
  zones,
  trawl_opening
)
}
\arguments{
\item{tutti_catch}{dataframe}

\item{tutti_operation}{dataframe}

\item{liste_station}{character}

\item{liste_dates}{dataframe}

\item{zones}{integer}

\item{trawl_opening}{numeric}
}
\value{
a list of 5 datatable
}
\description{
Regroupe toutes les fonctions pour la mise en forme et le calcul des indicateurs
group all the functions used to format the data and calculate the indicators
}
