% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_indice_computing.R
\name{indice_computing}
\alias{indice_computing}
\title{indice_computing}
\usage{
indice_computing(data_abun, data_biom)
}
\arguments{
\item{data_abun}{dataframe}

\item{data_biom}{dataframe}
}
\value{
a dataframe with the indicators
}
\description{
Calcul des indicateurs
calculate of the abundance, biomass, richness, shannon, simpson and
pielou indicators
}
