% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_leaflet_maker.R
\name{leaflet_maker}
\alias{leaflet_maker}
\title{leaflet_maker}
\usage{
leaflet_maker(df, shape)
}
\arguments{
\item{df}{dataframe}

\item{shape}{polygon}
}
\value{
a leaflet map
}
\description{
Creer la carte leaflet à partir du jeu de donnée mise en forme par la fonction preleaflet
create a leaflet map from a dataset created by the preleaflet function
}
