% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lineplot_creation.R
\name{lineplot_creation}
\alias{lineplot_creation}
\title{lineplot_creation}
\usage{
lineplot_creation(data, indicator, campagne)
}
\value{
The return value is a lineplot of the indicator chosen
}
\description{
A fct function that create lineplot using the dataframe
created by the function diversity_table
It represents the indicators from the diversity_table data in lineplot
with the standard error associated
The plot is just one survey
The indicators are (Abun, Biom, Richness, Shannon, Simpson)

 @param data dataframe
 @param indicator character
 @param campagne character
}
