% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lang.R
\name{mod_lang_ui}
\alias{mod_lang_ui}
\title{This script demonstrates how to use shiny.i18n Translator object
for live language change on the UI side with Shiny modules. Two key steps are:
(a) add `usei18n(i18n)` to UI
(b) use `update_lang` function to change the language in session}
\usage{
mod_lang_ui(id)
}
\description{
This script demonstrates how to use shiny.i18n Translator object
for live language change on the UI side with Shiny modules. Two key steps are:
(a) add `usei18n(i18n)` to UI
(b) use `update_lang` function to change the language in session
}
