% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_outlier_remove.R
\name{outlier_remove}
\alias{outlier_remove}
\title{outlier_remove}
\usage{
outlier_remove(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
The return value is a dataframe with the extrem value remove by the
 boxplot method
}
\description{
Remove survey lines with extreme values of the calculated
indicators
use of jehanne's method for her article, may need to be revised or removed
}
