% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_permanova_maker.R
\name{permanova_maker}
\alias{permanova_maker}
\title{permanova_maker}
\usage{
permanova_maker(data, formule, formule_bis, interaction, dist)
}
\arguments{
\item{data}{dataframe}

\item{formule}{expression from parse}

\item{formule_bis}{expression from parse}

\item{interaction}{Boolean}

\item{dist}{matrix or large list}
}
\value{
The return value is a list with the result of the permanova
with and without interaction
}
\description{
Fonction pour faire tourner les permanova en fonction de la formule creer par la fct ecriture_modele
L'intéraction est retirée automatiquement si non significative
}
