% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_probability_distribution.R
\name{probability_distribution}
\alias{probability_distribution}
\title{probability_distribution}
\usage{
probability_distribution(vector, law)
}
\value{
a vector density of a distribution probability
}
\description{
A fct function that take in input the type of distribution
and calculate the values of the probability distribution with the parameters
of a database
}
