% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_shapping.R
\name{table_shapping}
\alias{table_shapping}
\title{table_shapping}
\usage{
table_shapping(data)
}
\value{
The return value is a list of 5 dataframe
}
\description{
Creer les tables réduites à partir de la table complète
create data tables with only one explaineed variable with the explanatory
variables
}
