% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_text_list_translator.R
\name{text_list_translator}
\alias{text_list_translator}
\title{text_list_translator}
\usage{
text_list_translator()
}
\value{
The return value is a table that you can export in a repertory
and charge top choose the message you want to display
}
\description{
This function compiles all the long text present in the app
It create a table with a language per column
To add message, you write text into a character vector for each language
and then you assemble the vectors into a table
}
