% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_traitement_maker.R
\name{traitement_maker}
\alias{traitement_maker}
\title{traitement_maker}
\usage{
traitement_maker(data, station, dates, traitement)
}
\value{
The return value is a character corresponding to the qualitative
variable of BACI.
}
\description{
Creer la variable traitement en fonction des imformations
rentrées (stations et dates)
Create the traitement variable based on the information entered
(stations and dates)

 @param data dataframe
 @param station character
 @param dates Dates
 @param traitement character
}
