% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_traitement_maker_zones.R
\name{traitement_maker_zones}
\alias{traitement_maker_zones}
\title{traitement_maker_zones}
\usage{
traitement_maker_zones(data, liste_station, liste_dates, zones, traitement)
}
\arguments{
\item{data}{dataframe}

\item{liste_station}{character}

\item{liste_dates}{dataframe}

\item{zones}{integer}

\item{traitement}{character}
}
\value{
No return value
}
\description{
Repete la fonction traitement_maker pour chaque sous-zone/periode
Repeat the traitement_maker function for each zone
Currently we work with 1 zone but with the future dev there might be others
}
