% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_verification.R
\name{verification}
\alias{verification}
\title{verification}
\usage{
verification(name, format)
}
\arguments{
\item{name}{character}

\item{format}{character}
}
\value{
The return value is a booleen
}
\description{
A fct function that check if the columns name of a dataset
correspond to attempted names
So it check if a table is in the good shape
}
