% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tutti_function_traitement.R
\name{tutti_function_traitement}
\alias{tutti_function_traitement}
\title{tutti_function_traitement}
\usage{
tutti_function_traitement(
  tutti_catch,
  tutti_operation,
  liste_station,
  liste_dates,
  zones,
  trawl_opening
)
}
\arguments{
\item{tutti_catch}{dataframe}

\item{tutti_operation}{dataframe}

\item{liste_station}{character}

\item{liste_dates}{dataframe}

\item{zones}{integer}

\item{trawl_opening}{numeric}
}
\value{
The return value is a list of 2 dataframe
}
\description{
Table formatting using GranulatShiny format
use dplyr and tidyr
}
